annot: AnnotationObject should be the size as the text content
This makes the AnnotationObject to resize the frame to the size of
the text content, or else the text that is too big will be shown
outside the object's frame (rectangle). For some reason we need to
set the object as a "text frame" (mbTextFrame = true) or else it
will not allow to automatically adjust the frame to the text
content.

Change-Id: Ib2b9a7c07626fe8b0a2286cfd67c3bf37cf9af03
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169642
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins