tdf#92029 AutoCorrect: correct 's and 't in Dutch to apostrophe
“In Dutch (in Belgium as in the Netherlands), when you type "'s"
or "'t" in sentences like: "Ik ben 's morgens opgestaan", the
apostrophe is changed into an opening quote by the autocorrect
feature, which is incorrect and requires a manual correction.”

The bug reported with the suggested AutoCorrect extension
by Pierre François. The difference here is always to use the
de facto standard typographic apostrophe in the replacement.

Change-Id: I2ede612f72145abcca2bc1477ffae3cabc9e5641
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169645
Tested-by: Jenkins
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>