tdf#161817 Fix Autofill overlay when zoom is low
Prior to this patch, the calculation of the offset to draw the outline was incorrect, which caused the two lines of the outline to overlap.

This patch fixes this issue and makes the outline work at any zoom level.

Change-Id: I95d0b240a53beda872cdf6f0e7d75b48966217d0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169665
Reviewed-by: Rafael Lima <rafael.palma.lima@gmail.com>
Tested-by: Jenkins
(cherry picked from commit 7383298e141dcc3df23c8c849513b7782b0785e2)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170113