Related: tdf#161708 sw content controls: fix custom default of dropdown to PDF
Export the bugdoc to PDF without noticing that there is a content
control around some of the content and notice how the PDF export lacks
some words from the body text.

What happens is that content controls are exported to PDF as PDF forms
by default, and the selected option of a dropdown has to be an index, so
in case the text of the dropdown content control is not one of the
options, then the PDF will miss those words.

Fix the problem by inserting the text of the dropdown at the start if
there would be no valid index for it.

Also add a bit of padding around the rectangle of the content controls,
it turns out there is a default 1pt border in PDF, and this would lead
to a cut-off text at the end if we don't compensate for that border.

Change-Id: I99447894b320b42ad9ffe0d54d0190000621616b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169694
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
(cherry picked from commit b3e01bd3bdd37961057026ab7f2b57cbfab35d4a)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169670