tdf#161833 treat semi-transparent pixels as opaque
Limiting the contour wrapping polygon to only opaque pixels
causes clipping of any shadows or other semi-transaprent
areas in the image. So, instead of testing for fully opaque
pixels, treat pixels that are not fully transparent as opaque.

Also, tdf#161833 would cause semi-transparent pixels to be
treated as fully transparent pixels when calculating the wrap
contour for an image. To force the correct contour when loading
a document, force the contour to be recalculated by ignoring
the saved polygon if the contour is set to "recreate on edit".

Change-Id: Ibe256f54e1c82de30c2b7d5b92a69344b4a7ba10
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169745
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins
(cherry picked from commit 3da1d8a791cff6cf2dad29b64c9ad15b1a65d5b3)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169957
Reviewed-by: Ilmari Lauhakangas <ilmari.lauhakangas@libreoffice.org>
Tested-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>