NFC sw fmtcol.cxx: expression was always false
the NEW property by definition comes from pNewChgSet->GetChgSet(), and
the OLD property comes from GetAttrSet via GetItemIfSet.

It can never be "unless we set it", because by definition
    SfxPoolItem::areSame(pOld, pNew)
and therefore would not have gotten to this point of setting bContinue.

This logic has been there since initial import.
I assume the SfxPoolItem::areSame idea was added afterwards to
    // Avoid recursion (SetAttr!)
which obsoleted that clause.

-------------------------------------------------------------------
Note: Setting bContinue to false is simply wrong in the first place,
because just a single recalculation of a relative value
can halt the proliferation of other legitimate changes.

So, I'm rebuilding the logic one piece at a time
to more clearly document the logic changes I am making,
which should also help in case any of these cause a regression.

Change-Id: Ida57aaf519a82292137b5ed896763907db1971be
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169895
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins