tdf#161483 angle unit import for cell text rotation
Currently style:rotation-angle of table-cell-properties is imported
using ::sax::Converter::convertNumber. That imports only integer values
and ignores units. But ODF uses the data type 'angle'. That is a double
with optional units 'deg', 'rad' or 'grad'. A unit-less value is
treated as degrees.

The patch changes the import to use the new method
::sax::Converter::convertAngle(), that was introduced in commit
9f62c7a0f2333d1b7d179a43b3b0341dba7554a1.

Change-Id: I98c35a762e90b959a005f5bcdf810c5676cc8b6a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169891
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>