related tdf#134204 : notify change unless all properties are handled #1
This is an NFC-ish patch.

SetFormatAttr immediately calls a ClientNotify
about the single property that it has changed.
So if this was a direct property change,
or the set included only one property,
then there is no need to re-notify about this parent change.

The problem with the logic so far has been a failure
to notify about changes. In this patch, we had the opposite
problem of notifying unnecessarily, which isn't terrible.
The main benefit will be seen in a follow-up patch
which allows a single bContinue assignment per property.

Change-Id: I51d961f03910f0b8f50c6402894c5269f15bf00d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169897
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>