NFC tdf#134204: ensure style modify signaled if parent changed
If the parent changed, we have no idea how many properties might
be involved. Therefore, there is no way that anything could
legitimately say that we haven't actually changed/don't need to notify.

This is a No Functional Change.
bContinue was never set to false when bNewParent.
bNewParent and pNewChgSet are mutually exclusive,
and the only other time bContinue was changed was if pNewChgSet.

So this patch is just a simplification.
It more clearly shows that the notify should always happen,
and is helpful to demonstrate logic changes in the following patches.

Change-Id: Iabef5ba305c80d68655a0c880301ab32ffc90792
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169896
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins