related tdf#134204 : notify change unless all properties are handled #2
This patch is the biggest logic change in this patchset:
- no notify if single property (!pNewChgSet) didn't cause a change
    -bContinue can now be set false - avoiding a false notify
- multiple properties: don't block notify
    -by doing this, I can remove my earlier fix.
------------------------------------------------
Here is my logic:

    else if( pNewChgSet )
        bContinue = false
simply was incorrect logic. There is no way we can say that
no notification is needed simply because one property is unchanged.
AT BEST (which will be done in a follow-up patch)
we will need to track how many bContinue = false's there are
and see if they match the number of properties that were passed.
But sufficient should be to only block when Count == 1,
like pOldChgSet now does.
    bContinue = pNewChgSet->GetChgSet()->Count() > 1;

The variable pNewChgSet had been used earlier because it then
was checking if pNewChgSet->GetChgdSet...
But we can easily substitute pOldChgSet here, becoming
    else if (pOldChgSet)
        bContinue = pOldChgSet->GetChgSet()->Count() > 1;

There are three mutually exclusive code paths to consider:
   pOldChgSet, bNewParent, or just a single property
Since bNewParent overrides bContinue elsewhere, ignore it
and it also doesn't make sense to continue for a single property,
so this could simply become
    else
        bContinue = pOld... && pOld...->GetChgSet()->Count() > 1;
which is identical to the bContinue in the if clause,
and thus they could be combined.

Change-Id: Ifa5ba56226f4e77a00f3bc8089a8dcaaca2b91ea
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169898
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>