tdf#161853 Drop always null SvxIconChoiceCtrl_Impl::pHead
With

    Change-Id: I13ae34ffc51972bef746e53fcd46e65ea9fbf82c
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Jul 3 10:38:56 2024 +0200

        tdf#161853 Drop unused SvxIconChoiceCtrl_Impl::InitPredecessors

in place, it can be seen that `SvxIconChoiceCtrl_Impl::pHead`
is always nullptr as it's initialized to that and never
assigned a different value afterwards.

Therefore, drop it and all related code.

Drop the now unused `bKeepPredecessors` param
from `SvxIconChoiceCtrl_Impl::Arrange` and
`SvxIconChoiceCtrl_Impl::ImpArrange`.

(Use `git show --ignore-space change` to more clearly
see the "actual changes").

Change-Id: I44eceb41866bb6d99ccea032f98637a42dabc9ee
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169914
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins