tdf#152104 speed xls->ods convert part 2
We are doing a lot of work in ScAttrArray::SetPatternAreaImpl,
where we move the mvData vector of CellAttributeHolder around,
so inline and reduce some of the work.
The move operator= can be simplified, leaving some of the work
to the destructor (which may well be elided when moving
arrays of elements around).

Reduces time from 50s to 39s

Change-Id: I358f48fa30f785f7b3f221079db93ab0288a7f14
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169941
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>