tdf#101686 Writer textbox direction changes depending on zoom
Previously, DrawingLayer would apply RTL flags in the case of RTL text,
but it would not clear them in the case of LTR text.

In certain situations, it is possible for Writer to draw the contents of
a text box while the VCL output device is configured for RTL output. As
a result, whether DrawingLayer correctly drew LTR text as LTR would
depend on the order that text was rendered. This issue would manifest as
text boxes appearing to spontaneously mirror themselves while zooming
and scrolling through Writer documents containing RTL text.

This change updates the relevant portion of DrawingLayer to also clear
the RTL flags in the case of LTR text.

Change-Id: I3a75561d08cdf3ee6c74b8f6e62f75a1dfcf165e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169942
Tested-by: Jenkins
Reviewed-by: Jonathan Clark <jonathan@libreoffice.org>
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170928