tdf#161771 sw content controls: fix DOCX export of empty dropdown list item
Open the bugdoc, save as DOCX, try to open in Word: Word refuses to
open, saying that the file is corrupted.

Each dropdown item has a value and a display text, it seems it's OK to
omit the display text, but the value really should not be empty.

Fix the problem by first trying to copy the display text to the value if
the value would be empty; and if both are empty, then just omit the
dropdown item.

Note that the trick used at display text won't work here, omitting the
value attribute (instead of writing an empty one) is still invalid DOCX.

Change-Id: I4ae86aaf1a11cc8fd7c276634647f5737a9b04e4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170142
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins
(cherry picked from commit 62fb52cd43d7c0d41dd4e35a1c128947b6a14918)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170111
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>