jsdialog: enable interaction handler without LOKNotifier cool#9453
When pasting content with URL to remote content behind
basic authentication - we ask for user and password.
That dialog didn't have LOKNotifier set. It's new case
because we create the dialog controller and we setup the
window id = 0, later when we run the dialog we did correct
setup of LOKNotifier and id was changed to eg. 7.
Then client had out of sync identifier because we already sent
the old value. Setup the "default" notifier directly in
weld_dialog to avoid that.

Signed-off-by: Szymon Kłos <szymon.klos@collabora.com>
Change-Id: I7b6ceef722f7f9827b1fb41cc0229b0170693d1b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170159
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170192
Tested-by: Jenkins