tdf#161911 annot: fix for annotation with 0 size
We need to set the size to something non-zero or it will stay
zero even if we try to adapt the frame to the text content.

Additionally when we adapt the frame to text, we need to update
the annotation with the new size.

Change-Id: I4125f95fe6d0e55ab3b00a6a457cd2c9e04ec7c5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170260
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins
(cherry picked from commit b547e9f19de339531709c14931998125b9a41649)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170334
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>