tdf#161659: icon choice ctrl: Make use of the extra space
    commit f6ea343e6fb2dc3539823dee60c9c6f96fc16275
    Date:   Thu May 30 12:26:11 2024 +0200

        tdf#161030 Improve vertical Tab list display

increased the width for items in the vertical tab
control.
Make use of the new space also for items using an icon.

This causes them to be centered again (tdf#161659).

When using the kf5 VCL plugin with the default
Breeze style, this also causes the "New Document"
text of the last item in the "Insert" -> "Hyperlink"
dialog to no longer be cut off (see attachment 149544
from tdf#123623 for the previous behavior).
(As a side note, switching `SvxIconChoiceCtrlEntry::eTextMode`
from `SvxIconChoiceCtrlTextMode::Short`
to `SvxIconChoiceCtrlTextMode::Full` would generally
allow text to wrap and use multiple lines instead of
cutting off text at the end of the single line.)

Change-Id: I7ec2575bbeb895c034dfc9f2033cea3e3986305a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170262
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>
Reviewed-by: Rafael Lima <rafael.palma.lima@gmail.com>
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>