icon choice ctrl a11y: Don't explicitly unset selection
`SvtIconChoiceCtrl::SetCursor` already takes care of
unselecting all other entries when one is selected,
so there's no need to explicitly call
`SvtIconChoiceCtrl::SetNoSelection` first.

This can e.g. be tested by triggering the action
of the entries in the "Insert" -> "Hyperlink" dialog
in Writer from Accerciser when running LO with the
qt6 VCL plugin. The action causes the corresponding
entry to get selected.

Change-Id: Ic0c31d987b286095d58756fd92aacc662a723680
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170368
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>