icon choice ctrl a11y: Drop extra check
`AccessibleIconChoiceCtrl::getAccessibleChildCount`
returns the the number of elements in the
`SvxIconChoiceCtrl_Impl::maEntries` vector, and
`SvtIconChoiceCtrl->GetEntry` returns the
the element at the given index.

Given the index check above, a valid entry
should always be there, so drop the extra check
that would throw a RuntimeException
in `AccessibleIconChoiceCtrl::getAccessibleChild`.
The `pEntry` isn't otherwise used right here anyway.

Change-Id: I446b6d8d92780d484ebf5fc531bfa51a7c87cac7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170372
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>