tdf#161026 tdf#161355 icon choice ctrl: Draw icon + text at the end
Draw the entry's icon and text after drawing other things, to
make sure it is on top, i.e. always visible.

This shouldn't make any difference by itself yet as currently
only frames/borders are drawn after it, but that will change
with an upcoming commit to use the native drawing API for
taking care of at least the highlighting part.

Change-Id: I97cebdf2470ea1ab92e56642e4b4fd4255542994
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170340
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>