tdf#161026 tdf#161355 tdf#161501 icon choice ctrl: Natively draw mouse-hover
... feedback.

Instead of always using `SvxIconChoiceCtrl_Impl::DrawHighlightFrame`
to draw a frame around the currently mouse-hovered item in
the icon choice control/vertical tab page, use the native
drawing API to let the backends draw a `ControlType::TabItem`
with the `ControlState::ROLLOVER` state set.

This also aligns this more with the horizontal tabbar used in other dialogs
(s. `TabControl::ImplDrawItem`).

With this commit in place, the hovered vertical tab in
e.g. the "Insert" -> "Hyperlink" or "Format" -> "Page" vertical
tab dialogs is now displayed "properly" and in line with what
happens for non-vertical tab pages (like the "Format" -> "Paragraph"
one) on Windows or with the Breeze and Breeze Dark themes and
the kf5 VCL plugin on Linux.

Instead of just using the native drawing API for the mouse-hover
feedback, this could likely be extended to also take care of focus,
selection, etc. in a follow-up step.

Change-Id: I5437d18cb78e8da0736c9ab920c55b90af4d0e25
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170341
Reviewed-by: Rafael Lima <rafael.palma.lima@gmail.com>
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>