icon choice ctrl entry a11y: Use OAccessibleComponentHelper as base
Let `AccessibleIconChoiceCtrlEntry` derive from
`comphelper::OAccessibleComponentHelper` which
allows simplifying the class. It just requires
implementing the virtual `implGetBounds()` method
and provides implementations for the the other
location/size-related methods using that, so drop
the custom implementations doing the same.

This also makes it unnecessary to implement the
`XAccessibleEventBroadcaster` methods ourselves,
so drop the custom handling and
`AccessibleIconChoiceCtrlEntry::m_nClientId`
altogether as well.

Change-Id: Iaa14293a911ea7b199a21baf8c5c3cfb5075efcf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170380
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>