tdf#129244: Let ImpGraphic update its buffer for new pixel sizes
It used getVectorGraphicReplacement once,  getting a bitmap with pixel
size that depended on the default device DPI,  unrelated to the needed
bitmap size. It was not updated later, using the cached bitmap for any
requested size.

The DPI passed to convertPrimitive2DSequenceToBitmapEx  must represent
the relation  between the vector logic size  and the pixel size of the
generated bitmap, not the target device DPI.

This change introduces VectorGraphicData::getBitmap to obtain a bitmap
for any requested pixel size, and uses it every time in the introduced
ImpGraphic::updateBitmapFromVectorGraphic method every time it needs a
new size of the image.

Change-Id: Ife1305821fff28f0cd4070615ab032211a5d8c4f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170391
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>