tdf#161501 icon choice ctrl: Highlight full entry area
For background and borders/frames drawn to highlight
the currently selected/focused/mouse-hovered entry in
the icon choice control, use the entry's full area
and don't restrict the bottom edge to where the text
ends.

It's also unclear to me why 1 pixel would be subtracted
for the top and right edge when calculating the rect -
drop that as well.
(There is a line of a pixel or so in between the entries,
but since that actually doesn't belong to any of the
entries -- clicking there doesn't change the selected
entry -- also don't highlight it as part of one of the
entries.)

This addresses this (preexisting) issue mentioned in [1]:

> Maybe for a future patch... but the height from `CalcFocusRect`
> is too low. It barely covers the item and leaves gaps with
> respect to subsequent items.

[1] https://gerrit.libreoffice.org/c/core/+/170341/comment/57dc568d_1c0a8cf2/

Change-Id: I3369401c0bbdb77a054da9f57ac70f30f97c9cca
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170393
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Reviewed-by: Rafael Lima <rafael.palma.lima@gmail.com>
Tested-by: Jenkins