tdf#162002 DOCX import: ignore subsetted embedded fonts for editing
Open the bugdoc, some characters use a fallback font, some not, while
the same font is used in Word for all characters.

It turns out that the bugdoc has a subsetted embedded font, and the font
only has 8 glyphs, which explains the inconsistent rendering.

Fix the problem by ignoring subsetted fonts for the editing case: if a
font is subsetted then only use the embedded font for viewing. First
also had to fix up the tokenizer, as it had the wrong type for the
subsetted attribute, so we couldn't get its boolean value
(OOXMLStringValue doesn't override getInt(), so OOXMLValue::getInt()
returns 0 unconditionally).

Note that in theory we can have embedded fonts which are nominally
subsetted but in practice they have a "reasonable" amount of glyphs in
them, those are currently discarded, but should be probably still kept
in a follow-up change.

Change-Id: I3e6dfbb246d464fede64f183a3f244edc70e7b8a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170397
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>