tdf#128186 Enable native full screen mode on macOS
Commit c1452e73091412ba0bb72306329e1912df2ba513 disabled native full
screen mode on macOS so revert the minimal amount of pieces of that
commit. In theory, this should reintroduce one or more of the various
bugs identified in the above commit as well as in tdf#117872.

This approach purposely excludes all of the code that kept the native
and internal full screen modes in sync. This is because the fix for
tdf#161623 made native and internal full screen modes exist
independently of each other. Essentially, native full screen mode is
sort of treated like a window decoration and the internal full
screen mode is treated the same as it would if the native window was
not in full screen mode.

Change-Id: I45110d12648092cfb4d6a26d19dfcf956777f8ca
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170452
Tested-by: Jenkins
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>