vcl: Use nNewPos consistently
Use the `nNewPos` variable for assigning to
`mnCurPos` in `ToolBox::MouseButtonDown`
as well, rather than `i` which was the loop
variable above, and whose value was
assigned to `nNewPos` for the case that this
code path is entered.

Using `nNewPos` consistently makes a bit
more obvious that this is the same as in
the surrounding lines.

Change-Id: I5ce2da04310f21070931549a599248fdcad45317
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170489
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>