tdf#125051: rework and improve uitest
The test still fails if f7936b8ccd0ce9bea3a5e780443e4ae8f35b9da1
"tdf#125051: fix spellchecking on Calc" is reverted.
Change the test to avoid having the second dialog prompted when launching the spell dialog. The options dialog part is removed because it's not needed for testing the crash

Change-Id: I337376dd9ad0f5454534821d259589a9bd68af35
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170495
Tested-by: Jenkins
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>