Revert "tdf#159013 writerfilter: link to the correct first-page header"
This reverts commit 25.2 358ac0da827f908ce312c6d392a3d9102f36428e.
which was backported to 24.8

Although this patch made the situation "better than it was"
for most cases, it will inevitably make some documents worse.

Ultimately, this approach is wrong.
See tdf#158977's 158977_linkNonEmptyFirst.docx.

We can't only copy from the styles that "express" a H/F
(and UNO doesn't let us copy from styles that hide it).

A COPY appropach would need to ALWAYS import every header/footer
into a dedicated "header/footer repository" page style,
and every page style would need to copy its "expressed"
headers/footers from the repository to its page style.

That would mean our copy procedure would need to be very robust,
and all of this is way too big for a bug-fixer like me to consider.
I don't want to be the one on the hook expected to fix it properly.

Change-Id: I2af4a03a219daabf186473839ba444ff9797ec2f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170530
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>