tdf#158977: restore disabled rtfexport2 testFdo44176
Some unit tests were disabled with
    tdf#136472 adjust ooxml import to handle first header/footer

This unit test seems pointless currently,
but apparently back in 3.5 days it means something.
The point is that the header space on the two pages
is supposed to be identical, which measureably
means that the body area of the pages should be identical.

Adapted to try to keep the test alive and potentially relevant.

Change-Id: I9e03839a9a755ad604269ff46bdd556e8cae4393
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170533
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>