a11y: Clean up XAccessibleAction doc
The `XAccessibleAction::getAccessibleActionDescription`
and `XAcccessibleAction::doAccessibleAction` doc mentions
that an `IndexOutOfBoundsException` exception is thrown
when the index is invalid, which is indeed the expected
behaviour. However, they mentioned in addition that false
or an empty string is returned, which doesn't make sense
when exception gets thrown, so drop that.

Also slightly improve wording at 2 other places.

Change-Id: Idcc1fd669f32ff5a160cb35d2fbf6c2d23f65ab9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170567
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins