tdf#162002 DOCX import, font embed: only discard subset fonts with few glyphs
Commit a9f3c11375525a7708378dd3648febc40db1ad20 (tdf#162002 DOCX import:
ignore subsetted embedded fonts for editing, 2024-07-12) decided to
ignore all subsetted fonts for editing, improve this a little so we only
ignore subsetted fonts when they can't even provide an English alphabet
in any form (lowercase, uppercase).

This avoids the possible problem that a font is marked as subsetted but
it's good enough in practice and we would still throw it away for
editing.

Change-Id: I0bc0e14ffc0c039f029220991bd16d9e3254f059
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170570
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins