tdf#155447 gtk3 a11y: Use GtkBuilder ID as accessible ID
Set the GtkBuilder ID (the ID set in the .ui file)
as the accessible ID of the GtkWidget for welded
widgets in the gtk3 VCL plugin.

For non-welded widgets, that would happen
in `atk_object_wrapper_new`, similar to how

    Change-Id: I2281d15c38a229410469c29a14fdc4221aeb3e1e
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Jul 16 13:57:41 2024 +0200

        tdf#155447 qt a11y: Report accessible ID

implements it for the qt6 VCL plugin.

However, since native widgets are now used
almost everywhere for the gtk3 VCL plugin,
the implementation in `atk_object_wrapper_new`
does not take effect for most of the UI, including the
spell checking dialog mentioned in tdf#155447.
This commit makes sure an accessible ID is set for
such cases as well.

Using the GtkBuilder ID is also mentioned
in the `atk_object_set_accessible_id` doc [1]:

> Sets the accessible ID of the accessible. This is not meant to be
> presented to the user, but to be an ID which is stable over application
> development. Typically, this is the gtkbuilder ID. Such an ID will be
> available for instance to identify a given well-known accessible object
> for tailored screen reading, or for automatic regression testing.

GTK 4 currently doesn't have API to set an accessible ID.

[1] https://docs.gtk.org/atk/method.Object.set_accessible_id.html

Change-Id: I6061888f38aa40a0d3c4e680b7daa1649d5954e0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170586
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>