Resolves tdf#162018 Unexpected loss in focus in the "Manage Changes"
window when adding a change in the document (Track changes mode),
round 2

Addresses two situations where the list still jumps to the top:

Situation 1:
When an entry is added to the end of the changes list.

Situation 2:
After editing a comment for a change list entry with the 'Comment'
dialog.

Change-Id: I4a75e6bcba13fb9cc735cf78756fe736b9d4665d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170598
Reviewed-by: Jim Raykowski <raykowj@gmail.com>
Tested-by: Jenkins