Related: cool#9504 sw: don't invalidate num rules when pasting into bullets
Open the bugdoc, navigate to the last para of the first page, enable
bullets for that paragraph, paste a one-liner plain text string, observe
a 274 ms hang.

Commit dc3e022866893d8c0950f2269c0c8d61c24ed9bf (cool#9504 sw: don't
invalidate num rules when pasting into a non-num paragraph, 2024-07-15)
already fixed a not needed invalidation when pasting plain text at a
non-list paragraph, but here the insert point is part of a list, so we
still invalidate.

Improve this further to also not invalidate when we paste a paragraph
which is part of a list, but the numbering type is a bullet, which again
doesn't need an invalidation.

If the invalidation still happens, then the LOK case still calcs the
entire layout in one go, that's not yet fixed here.

Change-Id: I97c92aebb67ba8ca66b5820943dc1550d62bb467
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170606
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins
(cherry picked from commit 8af3bab06ed27a01df370bb5f79e5f15892e778d)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170649
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>