Correcting <aboveWord> and <belowWord> in the Danish locale.
Changed <aboveWord> from 'over' to 'ovenfor', and <belowWord> from 'under' to 'nedenfor'.

In Danish, 'over' is a correct translation if 'above' is followed by an object, for instance 'the caption is above[over] the table'. But <aboveWord> is used for cross-references where no object is following: 'see the figure above[ovenfor].'

Likewise for <belowWord>: one object is below[under] another object, but 'see below[nedenfor]'

Change-Id: Ib428342acf83c5961234aebbd50cb6a4a5d28d40
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170563
Tested-by: Jenkins
Reviewed-by: Eike Rathke <erack@redhat.com>