tdf#162268: Import - add zero width space after line break
If a paragraph ends with a line break Word applies the paragraph
properties to the last line. Writer extends the properties of the
las run.
The new line could be smaller or bigger as it would be in Word.
To prevent that a zero width space is added at the end of the
paragraph if the character properties of the run differ to
the properties of the paragraph.
Additionally attributes are not applied to the current font if the
portion is hidden.

Change-Id: I128d67bb621fb99d9db008cccbb93733f255294e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/173059
Reviewed-by: Thorsten Behrens <thorsten.behrens@allotropia.de>
Tested-by: allotropia jenkins <jenkins@allotropia.de>